+++ types/babel-types/package.json
@@ -2,8 +2,8 @@
     "private": true,
     "types": "index",
     "typesVersions": {
-        ">=3.7.0-0": {
-            "*": ["ts3.7/*"]
+        "<=3.6": {
+            "*": ["ts3.6/*"]
         }
     }
 }
