--- types/babel-types/ts3.7/index.d.ts
@@ -1,235 +1,0 @@
-export * from '../index';
-
-export function assertArrayExpression(node: object | null | undefined, opts?: object): asserts node is import('../index').ArrayExpression;
-export function assertAssignmentExpression(node: object | null | undefined, opts?: object): asserts node is import('../index').AssignmentExpression;
-export function assertBinaryExpression(node: object | null | undefined, opts?: object): asserts node is import('../index').BinaryExpression;
-export function assertDirective(node: object | null | undefined, opts?: object): asserts node is import('../index').Directive;
-export function assertDirectiveLiteral(node: object | null | undefined, opts?: object): asserts node is import('../index').DirectiveLiteral;
-export function assertBlockStatement(node: object | null | undefined, opts?: object): asserts node is import('../index').BlockStatement;
-export function assertBreakStatement(node: object | null | undefined, opts?: object): asserts node is import('../index').BreakStatement;
-export function assertCallExpression(node: object | null | undefined, opts?: object): asserts node is import('../index').CallExpression;
-export function assertCatchClause(node: object | null | undefined, opts?: object): asserts node is import('../index').CatchClause;
-export function assertConditionalExpression(node: object | null | undefined, opts?: object): asserts node is import('../index').ConditionalExpression;
-export function assertContinueStatement(node: object | null | undefined, opts?: object): asserts node is import('../index').ContinueStatement;
-export function assertDebuggerStatement(node: object | null | undefined, opts?: object): asserts node is import('../index').DebuggerStatement;
-export function assertDoWhileStatement(node: object | null | undefined, opts?: object): asserts node is import('../index').DoWhileStatement;
-export function assertEmptyStatement(node: object | null | undefined, opts?: object): asserts node is import('../index').EmptyStatement;
-export function assertExpressionStatement(node: object | null | undefined, opts?: object): asserts node is import('../index').ExpressionStatement;
-export function assertFile(node: object | null | undefined, opts?: object): asserts node is import('../index').File;
-export function assertForInStatement(node: object | null | undefined, opts?: object): asserts node is import('../index').ForInStatement;
-export function assertForStatement(node: object | null | undefined, opts?: object): asserts node is import('../index').ForStatement;
-export function assertFunctionDeclaration(node: object | null | undefined, opts?: object): asserts node is import('../index').FunctionDeclaration;
-export function assertFunctionExpression(node: object | null | undefined, opts?: object): asserts node is import('../index').FunctionExpression;
-export function assertIdentifier(node: object | null | undefined, opts?: object): asserts node is import('../index').Identifier;
-export function assertIfStatement(node: object | null | undefined, opts?: object): asserts node is import('../index').IfStatement;
-export function assertLabeledStatement(node: object | null | undefined, opts?: object): asserts node is import('../index').LabeledStatement;
-export function assertStringLiteral(node: object | null | undefined, opts?: object): asserts node is import('../index').StringLiteral;
-export function assertNumericLiteral(node: object | null | undefined, opts?: object): asserts node is import('../index').NumericLiteral;
-
-/** @deprecated Use `assertNumericLiteral` */
-export function assertNumberLiteral(node: object | null | undefined, opts?: object): asserts node is import('../index').NumericLiteral;
-export function assertNullLiteral(node: object | null | undefined, opts?: object): asserts node is import('../index').NullLiteral;
-export function assertBooleanLiteral(node: object | null | undefined, opts?: object): asserts node is import('../index').BooleanLiteral;
-export function assertRegExpLiteral(node: object | null | undefined, opts?: object): asserts node is import('../index').RegExpLiteral;
-
-/** @deprecated Use `assertRegExpLiteral` */
-export function assertRegexLiteral(node: object | null | undefined, opts?: object): asserts node is import('../index').RegExpLiteral;
-export function assertLogicalExpression(node: object | null | undefined, opts?: object): asserts node is import('../index').LogicalExpression;
-export function assertMemberExpression(node: object | null | undefined, opts?: object): asserts node is import('../index').MemberExpression;
-export function assertNewExpression(node: object | null | undefined, opts?: object): asserts node is import('../index').NewExpression;
-export function assertProgram(node: object | null | undefined, opts?: object): asserts node is import('../index').Program;
-export function assertObjectExpression(node: object | null | undefined, opts?: object): asserts node is import('../index').ObjectExpression;
-export function assertObjectMethod(node: object | null | undefined, opts?: object): asserts node is import('../index').ObjectMethod;
-export function assertObjectProperty(node: object | null | undefined, opts?: object): asserts node is import('../index').ObjectProperty;
-export function assertRestElement(node: object | null | undefined, opts?: object): asserts node is import('../index').RestElement;
-export function assertReturnStatement(node: object | null | undefined, opts?: object): asserts node is import('../index').ReturnStatement;
-export function assertSequenceExpression(node: object | null | undefined, opts?: object): asserts node is import('../index').SequenceExpression;
-export function assertSwitchCase(node: object | null | undefined, opts?: object): asserts node is import('../index').SwitchCase;
-export function assertSwitchStatement(node: object | null | undefined, opts?: object): asserts node is import('../index').SwitchStatement;
-export function assertThisExpression(node: object | null | undefined, opts?: object): asserts node is import('../index').ThisExpression;
-export function assertThrowStatement(node: object | null | undefined, opts?: object): asserts node is import('../index').ThrowStatement;
-export function assertTryStatement(node: object | null | undefined, opts?: object): asserts node is import('../index').TryStatement;
-export function assertUnaryExpression(node: object | null | undefined, opts?: object): asserts node is import('../index').UnaryExpression;
-export function assertUpdateExpression(node: object | null | undefined, opts?: object): asserts node is import('../index').UpdateExpression;
-export function assertVariableDeclaration(node: object | null | undefined, opts?: object): asserts node is import('../index').VariableDeclaration;
-export function assertVariableDeclarator(node: object | null | undefined, opts?: object): asserts node is import('../index').VariableDeclarator;
-export function assertWhileStatement(node: object | null | undefined, opts?: object): asserts node is import('../index').WhileStatement;
-export function assertWithStatement(node: object | null | undefined, opts?: object): asserts node is import('../index').WithStatement;
-export function assertAssignmentPattern(node: object | null | undefined, opts?: object): asserts node is import('../index').AssignmentPattern;
-export function assertArrayPattern(node: object | null | undefined, opts?: object): asserts node is import('../index').ArrayPattern;
-export function assertArrowFunctionExpression(node: object | null | undefined, opts?: object): asserts node is import('../index').ArrowFunctionExpression;
-export function assertClassBody(node: object | null | undefined, opts?: object): asserts node is import('../index').ClassBody;
-export function assertClassDeclaration(node: object | null | undefined, opts?: object): asserts node is import('../index').ClassDeclaration;
-export function assertClassExpression(node: object | null | undefined, opts?: object): asserts node is import('../index').ClassExpression;
-export function assertExportAllDeclaration(node: object | null | undefined, opts?: object): asserts node is import('../index').ExportAllDeclaration;
-export function assertExportDefaultDeclaration(node: object | null | undefined, opts?: object): asserts node is import('../index').ExportDefaultDeclaration;
-export function assertExportNamedDeclaration(node: object | null | undefined, opts?: object): asserts node is import('../index').ExportNamedDeclaration;
-export function assertExportSpecifier(node: object | null | undefined, opts?: object): asserts node is import('../index').ExportSpecifier;
-export function assertForOfStatement(node: object | null | undefined, opts?: object): asserts node is import('../index').ForOfStatement;
-export function assertImportDeclaration(node: object | null | undefined, opts?: object): asserts node is import('../index').ImportDeclaration;
-export function assertImportDefaultSpecifier(node: object | null | undefined, opts?: object): asserts node is import('../index').ImportDefaultSpecifier;
-export function assertImportNamespaceSpecifier(node: object | null | undefined, opts?: object): asserts node is import('../index').ImportNamespaceSpecifier;
-export function assertImportSpecifier(node: object | null | undefined, opts?: object): asserts node is import('../index').ImportSpecifier;
-export function assertMetaProperty(node: object | null | undefined, opts?: object): asserts node is import('../index').MetaProperty;
-export function assertClassMethod(node: object | null | undefined, opts?: object): asserts node is import('../index').ClassMethod;
-export function assertObjectPattern(node: object | null | undefined, opts?: object): asserts node is import('../index').ObjectPattern;
-export function assertSpreadElement(node: object | null | undefined, opts?: object): asserts node is import('../index').SpreadElement;
-export function assertSuper(node: object | null | undefined, opts?: object): asserts node is import('../index').Super;
-export function assertTaggedTemplateExpression(node: object | null | undefined, opts?: object): asserts node is import('../index').TaggedTemplateExpression;
-export function assertTemplateElement(node: object | null | undefined, opts?: object): asserts node is import('../index').TemplateElement;
-export function assertTemplateLiteral(node: object | null | undefined, opts?: object): asserts node is import('../index').TemplateLiteral;
-export function assertYieldExpression(node: object | null | undefined, opts?: object): asserts node is import('../index').YieldExpression;
-export function assertAnyTypeAnnotation(node: object | null | undefined, opts?: object): asserts node is import('../index').AnyTypeAnnotation;
-export function assertArrayTypeAnnotation(node: object | null | undefined, opts?: object): asserts node is import('../index').ArrayTypeAnnotation;
-export function assertBooleanTypeAnnotation(node: object | null | undefined, opts?: object): asserts node is import('../index').BooleanTypeAnnotation;
-export function assertBooleanLiteralTypeAnnotation(node: object | null | undefined, opts?: object): asserts node is import('../index').BooleanLiteralTypeAnnotation;
-export function assertNullLiteralTypeAnnotation(node: object | null | undefined, opts?: object): asserts node is import('../index').NullLiteralTypeAnnotation;
-export function assertClassImplements(node: object | null | undefined, opts?: object): asserts node is import('../index').ClassImplements;
-export function assertClassProperty(node: object | null | undefined, opts?: object): asserts node is import('../index').ClassProperty;
-export function assertDeclareClass(node: object | null | undefined, opts?: object): asserts node is import('../index').DeclareClass;
-export function assertDeclareFunction(node: object | null | undefined, opts?: object): asserts node is import('../index').DeclareFunction;
-export function assertDeclareInterface(node: object | null | undefined, opts?: object): asserts node is import('../index').DeclareInterface;
-export function assertDeclareModule(node: object | null | undefined, opts?: object): asserts node is import('../index').DeclareModule;
-export function assertDeclareTypeAlias(node: object | null | undefined, opts?: object): asserts node is import('../index').DeclareTypeAlias;
-export function assertDeclareVariable(node: object | null | undefined, opts?: object): asserts node is import('../index').DeclareVariable;
-export function assertExistentialTypeParam(node: object | null | undefined, opts?: object): asserts node is import('../index').ExistentialTypeParam;
-export function assertFunctionTypeAnnotation(node: object | null | undefined, opts?: object): asserts node is import('../index').FunctionTypeAnnotation;
-export function assertFunctionTypeParam(node: object | null | undefined, opts?: object): asserts node is import('../index').FunctionTypeParam;
-export function assertGenericTypeAnnotation(node: object | null | undefined, opts?: object): asserts node is import('../index').GenericTypeAnnotation;
-export function assertInterfaceExtends(node: object | null | undefined, opts?: object): asserts node is import('../index').InterfaceExtends;
-export function assertInterfaceDeclaration(node: object | null | undefined, opts?: object): asserts node is import('../index').InterfaceDeclaration;
-export function assertIntersectionTypeAnnotation(node: object | null | undefined, opts?: object): asserts node is import('../index').IntersectionTypeAnnotation;
-export function assertMixedTypeAnnotation(node: object | null | undefined, opts?: object): asserts node is import('../index').MixedTypeAnnotation;
-export function assertNullableTypeAnnotation(node: object | null | undefined, opts?: object): asserts node is import('../index').NullableTypeAnnotation;
-export function assertNumericLiteralTypeAnnotation(node: object | null | undefined, opts?: object): asserts node is import('../index').NumericLiteralTypeAnnotation;
-export function assertNumberTypeAnnotation(node: object | null | undefined, opts?: object): asserts node is import('../index').NumberTypeAnnotation;
-export function assertStringLiteralTypeAnnotation(node: object | null | undefined, opts?: object): asserts node is import('../index').StringLiteralTypeAnnotation;
-export function assertStringTypeAnnotation(node: object | null | undefined, opts?: object): asserts node is import('../index').StringTypeAnnotation;
-export function assertThisTypeAnnotation(node: object | null | undefined, opts?: object): asserts node is import('../index').ThisTypeAnnotation;
-export function assertTupleTypeAnnotation(node: object | null | undefined, opts?: object): asserts node is import('../index').TupleTypeAnnotation;
-export function assertTypeofTypeAnnotation(node: object | null | undefined, opts?: object): asserts node is import('../index').TypeofTypeAnnotation;
-export function assertTypeAlias(node: object | null | undefined, opts?: object): asserts node is import('../index').TypeAlias;
-export function assertTypeAnnotation(node: object | null | undefined, opts?: object): asserts node is import('../index').TypeAnnotation;
-export function assertTypeCastExpression(node: object | null | undefined, opts?: object): asserts node is import('../index').TypeCastExpression;
-export function assertTypeParameter(node: object | null | undefined, opts?: object): asserts node is import('../index').TypeParameter;
-export function assertTypeParameterDeclaration(node: object | null | undefined, opts?: object): asserts node is import('../index').TypeParameterDeclaration;
-export function assertTypeParameterInstantiation(node: object | null | undefined, opts?: object): asserts node is import('../index').TypeParameterInstantiation;
-export function assertObjectTypeAnnotation(node: object | null | undefined, opts?: object): asserts node is import('../index').ObjectTypeAnnotation;
-export function assertObjectTypeCallProperty(node: object | null | undefined, opts?: object): asserts node is import('../index').ObjectTypeCallProperty;
-export function assertObjectTypeIndexer(node: object | null | undefined, opts?: object): asserts node is import('../index').ObjectTypeIndexer;
-export function assertObjectTypeProperty(node: object | null | undefined, opts?: object): asserts node is import('../index').ObjectTypeProperty;
-export function assertQualifiedTypeIdentifier(node: object | null | undefined, opts?: object): asserts node is import('../index').QualifiedTypeIdentifier;
-export function assertUnionTypeAnnotation(node: object | null | undefined, opts?: object): asserts node is import('../index').UnionTypeAnnotation;
-export function assertVoidTypeAnnotation(node: object | null | undefined, opts?: object): asserts node is import('../index').VoidTypeAnnotation;
-export function assertJSXAttribute(node: object | null | undefined, opts?: object): asserts node is import('../index').JSXAttribute;
-export function assertJSXClosingElement(node: object | null | undefined, opts?: object): asserts node is import('../index').JSXClosingElement;
-export function assertJSXElement(node: object | null | undefined, opts?: object): asserts node is import('../index').JSXElement;
-export function assertJSXEmptyExpression(node: object | null | undefined, opts?: object): asserts node is import('../index').JSXEmptyExpression;
-export function assertJSXExpressionContainer(node: object | null | undefined, opts?: object): asserts node is import('../index').JSXExpressionContainer;
-export function assertJSXIdentifier(node: object | null | undefined, opts?: object): asserts node is import('../index').JSXIdentifier;
-export function assertJSXMemberExpression(node: object | null | undefined, opts?: object): asserts node is import('../index').JSXMemberExpression;
-export function assertJSXNamespacedName(node: object | null | undefined, opts?: object): asserts node is import('../index').JSXNamespacedName;
-export function assertJSXOpeningElement(node: object | null | undefined, opts?: object): asserts node is import('../index').JSXOpeningElement;
-export function assertJSXSpreadAttribute(node: object | null | undefined, opts?: object): asserts node is import('../index').JSXSpreadAttribute;
-export function assertJSXText(node: object | null | undefined, opts?: object): asserts node is import('../index').JSXText;
-export function assertNoop(node: object | null | undefined, opts?: object): asserts node is import('../index').Noop;
-export function assertParenthesizedExpression(node: object | null | undefined, opts?: object): asserts node is import('../index').ParenthesizedExpression;
-export function assertAwaitExpression(node: object | null | undefined, opts?: object): asserts node is import('../index').AwaitExpression;
-export function assertBindExpression(node: object | null | undefined, opts?: object): asserts node is import('../index').BindExpression;
-export function assertDecorator(node: object | null | undefined, opts?: object): asserts node is import('../index').Decorator;
-export function assertDoExpression(node: object | null | undefined, opts?: object): asserts node is import('../index').DoExpression;
-export function assertExportDefaultSpecifier(node: object | null | undefined, opts?: object): asserts node is import('../index').ExportDefaultSpecifier;
-export function assertExportNamespaceSpecifier(node: object | null | undefined, opts?: object): asserts node is import('../index').ExportNamespaceSpecifier;
-export function assertRestProperty(node: object | null | undefined, opts?: object): asserts node is import('../index').RestProperty;
-export function assertSpreadProperty(node: object | null | undefined, opts?: object): asserts node is import('../index').SpreadProperty;
-export function assertExpression(node: object | null | undefined, opts?: object): asserts node is import('../index').Expression;
-export function assertBinary(node: object | null | undefined, opts?: object): asserts node is import('../index').Binary;
-export function assertScopable(node: object | null | undefined, opts?: object): asserts node is import('../index').Scopable;
-export function assertBlockParent(node: object | null | undefined, opts?: object): asserts node is import('../index').BlockParent;
-export function assertBlock(node: object | null | undefined, opts?: object): asserts node is import('../index').Block;
-export function assertStatement(node: object | null | undefined, opts?: object): asserts node is import('../index').Statement;
-export function assertTerminatorless(node: object | null | undefined, opts?: object): asserts node is import('../index').Terminatorless;
-export function assertCompletionStatement(node: object | null | undefined, opts?: object): asserts node is import('../index').CompletionStatement;
-export function assertConditional(node: object | null | undefined, opts?: object): asserts node is import('../index').Conditional;
-export function assertLoop(node: object | null | undefined, opts?: object): asserts node is import('../index').Loop;
-export function assertWhile(node: object | null | undefined, opts?: object): asserts node is import('../index').While;
-export function assertExpressionWrapper(node: object | null | undefined, opts?: object): asserts node is import('../index').ExpressionWrapper;
-export function assertFor(node: object | null | undefined, opts?: object): asserts node is import('../index').For;
-export function assertForXStatement(node: object | null | undefined, opts?: object): asserts node is import('../index').ForXStatement;
-
-export function assertFunction(node: object | null | undefined, opts?: object): asserts node is import('../index').Function;
-export function assertFunctionParent(node: object | null | undefined, opts?: object): asserts node is import('../index').FunctionParent;
-export function assertPureish(node: object | null | undefined, opts?: object): asserts node is import('../index').Pureish;
-export function assertDeclaration(node: object | null | undefined, opts?: object): asserts node is import('../index').Declaration;
-export function assertLVal(node: object | null | undefined, opts?: object): asserts node is import('../index').LVal;
-export function assertLiteral(node: object | null | undefined, opts?: object): asserts node is import('../index').Literal;
-export function assertImmutable(node: object | null | undefined, opts?: object): asserts node is import('../index').Immutable;
-export function assertUserWhitespacable(node: object | null | undefined, opts?: object): asserts node is import('../index').UserWhitespacable;
-export function assertMethod(node: object | null | undefined, opts?: object): asserts node is import('../index').Method;
-export function assertObjectMember(node: object | null | undefined, opts?: object): asserts node is import('../index').ObjectMember;
-export function assertProperty(node: object | null | undefined, opts?: object): asserts node is import('../index').Property;
-export function assertUnaryLike(node: object | null | undefined, opts?: object): asserts node is import('../index').UnaryLike;
-export function assertPattern(node: object | null | undefined, opts?: object): asserts node is import('../index').Pattern;
-export function assertClass(node: object | null | undefined, opts?: object): asserts node is import('../index').Class;
-export function assertModuleDeclaration(node: object | null | undefined, opts?: object): asserts node is import('../index').ModuleDeclaration;
-export function assertExportDeclaration(node: object | null | undefined, opts?: object): asserts node is import('../index').ExportDeclaration;
-export function assertModuleSpecifier(node: object | null | undefined, opts?: object): asserts node is import('../index').ModuleSpecifier;
-export function assertFlow(node: object | null | undefined, opts?: object): asserts node is import('../index').Flow;
-export function assertFlowBaseAnnotation(node: object | null | undefined, opts?: object): asserts node is import('../index').FlowBaseAnnotation;
-export function assertFlowDeclaration(node: object | null | undefined, opts?: object): asserts node is import('../index').FlowDeclaration;
-export function assertJSX(node: object | null | undefined, opts?: object): asserts node is import('../index').JSX;
-
-export function assertTSAnyKeyword(node: object | null | undefined, opts?: object): asserts node is import('../index').TSAnyKeyword;
-export function assertTSArrayType(node: object | null | undefined, opts?: object): asserts node is import('../index').TSArrayType;
-export function assertTSAsExpression(node: object | null | undefined, opts?: object): asserts node is import('../index').TSAsExpression;
-export function assertTSBooleanKeyword(node: object | null | undefined, opts?: object): asserts node is import('../index').TSBooleanKeyword;
-export function assertTSCallSignatureDeclaration(node: object | null | undefined, opts?: object): asserts node is import('../index').TSCallSignatureDeclaration;
-export function assertTSConstructSignatureDeclaration(node: object | null | undefined, opts?: object): asserts node is import('../index').TSConstructSignatureDeclaration;
-export function assertTSConstructorType(node: object | null | undefined, opts?: object): asserts node is import('../index').TSConstructorType;
-export function assertTSDeclareFunction(node: object | null | undefined, opts?: object): asserts node is import('../index').TSDeclareFunction;
-export function assertTSDeclareMethod(node: object | null | undefined, opts?: object): asserts node is import('../index').TSDeclareMethod;
-export function assertTSEnumDeclaration(node: object | null | undefined, opts?: object): asserts node is import('../index').TSEnumDeclaration;
-export function assertTSEnumMember(node: object | null | undefined, opts?: object): asserts node is import('../index').TSEnumMember;
-export function assertTSExportAssignment(node: object | null | undefined, opts?: object): asserts node is import('../index').TSExportAssignment;
-export function assertTSExpressionWithTypeArguments(node: object | null | undefined, opts?: object): asserts node is import('../index').TSExpressionWithTypeArguments;
-export function assertTSExternalModuleReference(node: object | null | undefined, opts?: object): asserts node is import('../index').TSExternalModuleReference;
-export function assertTSFunctionType(node: object | null | undefined, opts?: object): asserts node is import('../index').TSFunctionType;
-export function assertTSImportEqualsDeclaration(node: object | null | undefined, opts?: object): asserts node is import('../index').TSImportEqualsDeclaration;
-export function assertTSIndexSignature(node: object | null | undefined, opts?: object): asserts node is import('../index').TSIndexSignature;
-export function assertTSIndexedAccessType(node: object | null | undefined, opts?: object): asserts node is import('../index').TSIndexedAccessType;
-export function assertTSInterfaceBody(node: object | null | undefined, opts?: object): asserts node is import('../index').TSInterfaceBody;
-export function assertTSInterfaceDeclaration(node: object | null | undefined, opts?: object): asserts node is import('../index').TSInterfaceDeclaration;
-export function assertTSIntersectionType(node: object | null | undefined, opts?: object): asserts node is import('../index').TSIntersectionType;
-export function assertTSLiteralType(node: object | null | undefined, opts?: object): asserts node is import('../index').TSLiteralType;
-export function assertTSMappedType(node: object | null | undefined, opts?: object): asserts node is import('../index').TSMappedType;
-export function assertTSMethodSignature(node: object | null | undefined, opts?: object): asserts node is import('../index').TSMethodSignature;
-export function assertTSModuleBlock(node: object | null | undefined, opts?: object): asserts node is import('../index').TSModuleBlock;
-export function assertTSModuleDeclaration(node: object | null | undefined, opts?: object): asserts node is import('../index').TSModuleDeclaration;
-export function assertTSNamespaceExportDeclaration(node: object | null | undefined, opts?: object): asserts node is import('../index').TSNamespaceExportDeclaration;
-export function assertTSNeverKeyword(node: object | null | undefined, opts?: object): asserts node is import('../index').TSNeverKeyword;
-export function assertTSNonNullExpression(node: object | null | undefined, opts?: object): asserts node is import('../index').TSNonNullExpression;
-export function assertTSNullKeyword(node: object | null | undefined, opts?: object): asserts node is import('../index').TSNullKeyword;
-export function assertTSNumberKeyword(node: object | null | undefined, opts?: object): asserts node is import('../index').TSNumberKeyword;
-export function assertTSObjectKeyword(node: object | null | undefined, opts?: object): asserts node is import('../index').TSObjectKeyword;
-export function assertTSParameterProperty(node: object | null | undefined, opts?: object): asserts node is import('../index').TSParameterProperty;
-export function assertTSParenthesizedType(node: object | null | undefined, opts?: object): asserts node is import('../index').TSParenthesizedType;
-export function assertTSPropertySignature(node: object | null | undefined, opts?: object): asserts node is import('../index').TSPropertySignature;
-export function assertTSQualifiedName(node: object | null | undefined, opts?: object): asserts node is import('../index').TSQualifiedName;
-export function assertTSStringKeyword(node: object | null | undefined, opts?: object): asserts node is import('../index').TSStringKeyword;
-export function assertTSSymbolKeyword(node: object | null | undefined, opts?: object): asserts node is import('../index').TSSymbolKeyword;
-export function assertTSThisType(node: object | null | undefined, opts?: object): asserts node is import('../index').TSThisType;
-export function assertTSTupleType(node: object | null | undefined, opts?: object): asserts node is import('../index').TSTupleType;
-export function assertTSTypeAliasDeclaration(node: object | null | undefined, opts?: object): asserts node is import('../index').TSTypeAliasDeclaration;
-export function assertTSTypeAnnotation(node: object | null | undefined, opts?: object): asserts node is import('../index').TSTypeAnnotation;
-export function assertTSTypeAssertion(node: object | null | undefined, opts?: object): asserts node is import('../index').TSTypeAssertion;
-export function assertTSTypeLiteral(node: object | null | undefined, opts?: object): asserts node is import('../index').TSTypeLiteral;
-export function assertTSTypeOperator(node: object | null | undefined, opts?: object): asserts node is import('../index').TSTypeOperator;
-export function assertTSTypeParameter(node: object | null | undefined, opts?: object): asserts node is import('../index').TSTypeParameter;
-export function assertTSTypeParameterDeclaration(node: object | null | undefined, opts?: object): asserts node is import('../index').TSTypeParameterDeclaration;
-export function assertTSTypeParameterInstantiation(node: object | null | undefined, opts?: object): asserts node is import('../index').TSTypeParameterInstantiation;
-export function assertTSTypePredicate(node: object | null | undefined, opts?: object): asserts node is import('../index').TSTypePredicate;
-export function assertTSTypeQuery(node: object | null | undefined, opts?: object): asserts node is import('../index').TSTypeQuery;
-export function assertTSTypeReference(node: object | null | undefined, opts?: object): asserts node is import('../index').TSTypeReference;
-export function assertTSUndefinedKeyword(node: object | null | undefined, opts?: object): asserts node is import('../index').TSUndefinedKeyword;
-export function assertTSUnionType(node: object | null | undefined, opts?: object): asserts node is import('../index').TSUnionType;
-export function assertTSVoidKeyword(node: object | null | undefined, opts?: object): asserts node is import('../index').TSVoidKeyword;
